Translators module + shared source folder
Now all the code shared across different Python processes
(in particular, the Experiment-wielding script and th helper)
is supposed to go to ./shared/ . The module routes.py is a notable
exception from this rule, cause it's super special.

grid.py is a part of the shared code, so it has been moved there.

And the first module to be added there is translators.py -
collection of the functions which translate data structures to
strings and vice versa. Because the protocol of the conversion
must be the same tor transmiters and receivers, it is very critical
for this module to not be distributed across the codebase.

No debug yet.

	modified:   lowPassEvolvability/experiment.py
	modified:   lowPassEvolvability/helpers/helper.py
	modified:   lowPassEvolvability/helpers/staticEvsDynamicCe.py
	modified:   lowPassEvolvability/runExperiment.py
	new file:   lowPassEvolvability/shared/__init__.py
	renamed:    lowPassEvolvability/grid.py -> lowPassEvolvability/shared/grid.py
	new file:   lowPassEvolvability/shared/translators.py