Inheritance architecture reorganized
Experiment, Helper and staticEvsDynamicCeExperiment are now truly
abstract base classes. Exposure of Experiment's methods is
changed and the class got better documentation. A bunch of stuff
which was written statically in the code is now a part of the
modifiable environment. Implementation of Experiment's final
children is now done at the executables.

No debug yet.

	modified:   env/pbsVacc.py
	modified:   experiment.py
	modified:   helpers/helper.py
	renamed:    runExperiment.py -> uniformSamplingLPE.py
	modified:   staticEvsDynamicCeExperiment.py